Refactors ATT parsing into a dedicated class
Moves the ATT packet parsing logic from the `BTContext` and
`BTSessionDissector` classes into a dedicated `ATTParser` class.
This improves code organization and reusability. Also, it improves
the way `BTSessionDissector` is initialized by creating a
`dissector` method to instantiate it only if needed.

Updates Command class to use contexts attribute instead of context
to follow the same data structure logic in the code.